Renamed the NyTimesArticleService -> NyTimesArticleApi. Moved its interface to the same file. Moved the retrofit interface definition to its own file and setup DI. Removed unnecessary enum and interface dynamic and hardcoded the query.
