+++ app/src/main/java/com/nytimesmostviewedarticles/Constants.kt
@@ -1,7 +1,5 @@
 package com.nytimesmostviewedarticles
 
-import com.nytimesmostviewedarticles.network.NyTimesArticlePeriod
-
 object Constants {
     // navigation keys
     const val MAIN_SCREEN = "Main Screen Key"
@@ -9,9 +7,6 @@ object Constants {
 
     // NY times articles API key
     const val API_KEY = "nKLx7rAx32IP9qsHdVcachu1zsGEcWu7"
-
-    // ViewModel Constants
-    val DEFAULT_PERIOD_ENUM = NyTimesArticlePeriod.WEEK
 
     // UI values
     const val DEFAULT_DETAIL_IMAGE_WIDTH = 440
