+++ app/src/main/java/com/nytimesmostviewedarticles/di/NyTimesArticleServiceModule.kt
@@ -1,0 +1,29 @@
+package com.nytimesmostviewedarticles.di
+
+import com.nytimesmostviewedarticles.network.NyTimesApiService
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+@InstallIn(SingletonComponent::class)
+@Module
+class NyTimesArticleServiceModule {
+    @Provides
+    fun providesNyTimesArticleService(): NyTimesApiService {
+        val baseUrl = "https://api.nytimes.com/svc/mostpopular/v2/viewed/"
+
+        val moshi = Moshi.Builder()
+            .add(KotlinJsonAdapterFactory())
+            .build()
+
+        return Retrofit.Builder()
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .baseUrl(baseUrl)
+            .build().create(NyTimesApiService::class.java)
+    }
+}
